refactor(test): share a common pattern to setup tests to detect or not changes
