+++ packages/skott/test/integration/__fixtures__/some-dir/some-other-dir/index.js
@@ -1,0 +1,1 @@
+console.log(1);
